Module rustypub::core::object

Structs

Application

Group

Link

A Link is an indirect, qualified reference to a resource identified by a URL. The fundamental model for links is established by RFC5988. Many of the properties defined by the Activity Vocabulary allow values that are either instances of Object or Link. When a Link is used, it establishes a qualified relation connecting the subject (the containing object) to the resource identified by the href. Properties of the Link are properties of the reference as opposed to properties of the resource.

LinkBuilder

Builder for Link.

Object

The Object is the primary base type for the Activity Streams vocabulary. In addition to having a global identifier (expressed as an absolute IRI using the id property) and an “object type” (expressed using the type property), all instances of the Object type share a common set of properties normatively defined by the Activity Vocabulary. These include: attachment | attributedTo | audience | content | context | contentMap | name | nameMap | endTime | generator | icon | image | inReplyTo | location | preview | published | replies | startTime | summary | summaryMap | tag | updated | url | to | bto | cc | bcc | mediaType | duration All properties are optional (including the id and type).

ObjectBuilder

Builder for Object.

Organization

Person

Preview

Identifies an entity that provides a preview of this object.

PreviewBuilder

Builder for Preview.

Service

Enums

AttributedTo

LinkBuilderError

Error type for LinkBuilder

ObjectBuilderError

Error type for ObjectBuilder

PreviewBuilderError

Error type for PreviewBuilder