Struct rustypub::core::Activity

pub struct Activity {
    pub actor: Option<Actor>,
    pub object: Option<Object<Null>>,
    pub target: Option<Object<Null>>,
    pub result: Option<String>,
    pub origin: Option<String>,
    pub instrument: Option<String>,
    /* private fields */
}

An Activity is a subtype of Object that describes some form of action that may happen, is currently happening, or has already happened. The Activity type itself serves as an abstract base type for all types of activities. It is important to note that the Activity type itself does not carry any specific semantics about the kind of action being taken.

Fields
actor: Option<Actor>object: Option<Object<Null>>target: Option<Object<Null>>result: Option<String>origin: Option<String>instrument: Option<String>